LF-14096 souphttpsrc: set seekable to TRUE only when Accept-Ranges is "bytes"
The original design set seekable to TRUE when souphttpsrc got new size,
And reset it to FALSE when get "none" Accept-Ranges in http header,
But some stream server will not feedback Accept-Ranges in http header
which means the seekable flag is not correct. Therefore, we need only
set it to TRUE when get valid Accept-Ranges in header.

upstream status: submitted
https://gitlab.freedesktop.org/gstreamer/gstreamer/-/merge_requests/8505