meson: Fix libsoup lookup when soup2 is installed
We were first looking for soup3 and we allowed fallback. That means that
if soup2 was installed it would still configure the fallback.

Fix by first looking for soup2 which has no fallback (libsoup.wrap
provides only libsoup-3.0) and look for soup3 only if soup2 wasn't
found (either not installed or soup-version=3).

Even if soup-lookup-dep=true, soup dependency is only required if at
least one plugin that uses it is enabled.

Part-of: <https://gitlab.freedesktop.org/gstreamer/gstreamer/-/merge_requests/9775>